+++ src/main/java/gdsc/binaryho/imhere/core/auth/application/AuthService.java
@@ -1,16 +1,14 @@
 package gdsc.binaryho.imhere.core.auth.application;
 
-import gdsc.binaryho.imhere.core.auth.model.response.SignInRequestValidationResult;
-import gdsc.binaryho.imhere.core.auth.model.request.SignInRequest;
 import gdsc.binaryho.imhere.core.auth.exception.DuplicateEmailException;
 import gdsc.binaryho.imhere.core.auth.exception.MemberNotFoundException;
 import gdsc.binaryho.imhere.core.auth.exception.PasswordFormatMismatchException;
 import gdsc.binaryho.imhere.core.auth.exception.PasswordIncorrectException;
-import gdsc.binaryho.imhere.core.auth.util.AuthenticationHelper;
+import gdsc.binaryho.imhere.core.auth.model.request.SignInRequest;
+import gdsc.binaryho.imhere.core.auth.model.response.SignInRequestValidationResult;
 import gdsc.binaryho.imhere.core.member.Member;
 import gdsc.binaryho.imhere.core.member.MemberRepository;
 import gdsc.binaryho.imhere.core.member.Role;
-import gdsc.binaryho.imhere.core.member.model.request.RoleChangeRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -22,7 +20,6 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class AuthService {
 
-    private final AuthenticationHelper authenticationHelper;
     private final MemberRepository memberRepository;
     private final BCryptPasswordEncoder bCryptPasswordEncoder;
 
@@ -66,19 +63,5 @@ public class AuthService {
         if (!bCryptPasswordEncoder.matches(rawPassword, encodedPassword)) {
             throw PasswordIncorrectException.EXCEPTION;
         }
-    }
-
-    @Transactional
-    public void memberRoleChange(RoleChangeRequest roleChangeRequest, String univId) {
-        authenticationHelper.verifyMemberHasAdminRole();
-
-        Member targetMember = memberRepository.findByUnivId(univId)
-            .orElseThrow(() -> MemberNotFoundException.EXCEPTION);
-
-        Role newRole = Role.valueOf(roleChangeRequest.getRole());
-        targetMember.setRole(newRole);
-
-        log.info("[권한 변경] " + univId + "의 권한이 {} 로 변경. ({})",
-            () -> roleChangeRequest.getRole(), () -> authenticationHelper.getCurrentMember().getUnivId());
     }
 }
